+++ include/IeepromMPU.h
@@ -1,20 +1,22 @@
 #pragma once
 
+#include <stdint.h>
+
 class IeepromMPU {
     public:
-        virtual int getXGyroOffset() = 0;
-        virtual int getYGyroOffset() = 0;
-        virtual int getZGyroOffset() = 0;
-        virtual int getXAccOffset() = 0;
-        virtual int getYAccOffset() = 0;
-        virtual int getZAccOffset() = 0;
+        virtual int16_t getXGyroOffset() = 0;
+        virtual int16_t getYGyroOffset() = 0;
+        virtual int16_t getZGyroOffset() = 0;
+        virtual int16_t getXAccOffset() = 0;
+        virtual int16_t getYAccOffset() = 0;
+        virtual int16_t getZAccOffset() = 0;
 
-        virtual void setXGyroOffset(int offset) = 0;
-        virtual void setYGyroOffset(int offset) = 0;
-        virtual void setZGyroOffset(int offset) = 0;
-        virtual void setXAccOffset(int offset) = 0;
-        virtual void setYAccOffset(int offset) = 0;
-        virtual void setZAccOffset(int offset) = 0;
+        virtual void setXGyroOffset(int16_t offset) = 0;
+        virtual void setYGyroOffset(int16_t offset) = 0;
+        virtual void setZGyroOffset(int16_t offset) = 0;
+        virtual void setXAccOffset(int16_t offset) = 0;
+        virtual void setYAccOffset(int16_t offset) = 0;
+        virtual void setZAccOffset(int16_t offset) = 0;
 
         virtual bool getCalibrationFlag() = 0;
         virtual void setCalibrationFlag() = 0;
